refactor: fix validation to using methods with auth token

+++ src/clients/clients.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common'
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common'
 import { CreateClientDto } from './dto/create-client.dto'
 import { UpdateClientDto } from './dto/update-client.dto'
 import { ClientEntity } from './entities/client.entity'
 import { InjectRepository } from '@nestjs/typeorm'
 import { ILike, Repository } from 'typeorm'
+import { userInfo } from 'os'
 
 @Injectable()
 export class ClientsService {
@@ -37,17 +38,45 @@ export class ClientsService {
     return serachClients
   }
 
-  async findOne(user_id: string) {
-    return await this.clientRepository.findOneBy({ user_id })
+  async findOne(id:string, user_id: string) {
+    return await this.clientRepository.findOne({ 
+      where: {
+        id: id,
+        user_id: user_id
+      }
+     })
   }
 
-  async updateClient(id: string, updateClientDto: UpdateClientDto) {
-    await this.clientRepository.update(id, updateClientDto)
-    const updateClient = await this.clientRepository.findOne({ where: { id } })
-    return updateClient
+  async updateClient(id: string, updateClientDto: UpdateClientDto, user_id: string) {
+    const updateClient = await this.clientRepository.findOne({
+      where: {
+        id: id
+      }
+    })
+
+    if (!updateClient) throw new NotFoundException('Cliente não encontrado')   
+
+    if (updateClient?.user_id !== user_id) {
+      throw new ForbiddenException("Voce nao tem permissao para atualizar esse cliente")
+    }
+
+    return await this.clientRepository.update(id, updateClientDto)
+
   }
 
-  async remove(id: string) {
+  async remove(id: string, user_id: string) {
+    const removeClient = await this.clientRepository.findOne({
+      where: {
+        id: id
+      }
+    })
+
+    if (!removeClient) throw new NotFoundException('Cliente não encontrado')  
+
+    if (removeClient?.user_id !== user_id) {
+      throw new ForbiddenException("Voce nao tem permissao para deletar esse cliente")
+    }
+
     return await this.clientRepository.delete(id)
   }
 }
